Moved builder application modal button to the cohort main page, cleaned up builder applicacions screen of redundant texts and restricted github usernames to alphanumeric characters
