+++ packages/nextjs/app/(routes)/cohort/[cohortAddress]/_components/ApplicationModal.tsx
@@ -1,62 +1,38 @@
-"use client";
-
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Application, Builder } from "@prisma/client";
 import axios from "axios";
-import { Plus } from "lucide-react";
 import { useForm } from "react-hook-form";
 import { useSignMessage } from "wagmi";
 import { useAccount } from "wagmi";
 import * as z from "zod";
-import { RainbowKitCustomConnectButton } from "~~/components/scaffold-eth";
 import { notification } from "~~/utils/scaffold-eth";
+
+interface ApplicationModalProps {
+  cohortAddress: string;
+  onApplicationSuccess?: () => void;
+}
 
 const CreateApplicationSchema = z.object({
   description: z.string().min(10, "Description must be at least 10 characters"),
-  githubUsername: z.string().optional(),
+  githubUsername: z
+    .string()
+    .max(30, "GitHub username cannot exceed 30 characters")
+    .regex(
+      /^(?!-)[a-zA-Z0-9-]+(?<!-)$/,
+      "GitHub username may only contain alphanumeric characters or single hyphens, and cannot begin or end with a hyphen",
+    )
+    .optional(),
 });
 
-interface ApplicationListProps {
-  applications: Application[];
-  builders: Builder[];
-  adminAddresses: string[];
-  cohortAddress: string;
-}
-
-export const ApplicationList = ({ applications, cohortAddress, builders, adminAddresses }: ApplicationListProps) => {
-  const router = useRouter();
+export const ApplicationModal = ({ cohortAddress, onApplicationSuccess }: ApplicationModalProps) => {
   const { address } = useAccount();
-
+  const router = useRouter();
   const [isPending, setIsPending] = useState(false);
   const [isSuccess, setIsSuccess] = useState(false);
   const [formValues, setFormValues] = useState<z.infer<typeof CreateApplicationSchema> | null>(null);
-  const [userApplications, setUserApplications] = useState<Application[]>([]);
 
   const { data: signature, signMessage, isSuccess: isSignatureSuccess } = useSignMessage();
-
-  const isBuilderorAdmin = () => {
-    if (!address) return false;
-    const isBuilder = builders.some(builder => builder.address.toLowerCase() === address.toLowerCase());
-    const isAdmin = adminAddresses.some(admin => admin.toLowerCase() === address.toLowerCase());
-    return isBuilder || isAdmin;
-  };
-
-  useEffect(() => {
-    if (address && applications) {
-      const filtered = applications.filter(app => app.address.toLowerCase() === address.toLowerCase());
-      setUserApplications(filtered);
-    }
-  }, [address, applications]);
-
-  const canApply = () => {
-    if (!userApplications.length) return true;
-
-    const hasPendingOrApproved = userApplications.some(app => app.status === "PENDING" || app.status === "APPROVED");
-
-    return !hasPendingOrApproved;
-  };
 
   const form = useForm<z.infer<typeof CreateApplicationSchema>>({
     resolver: zodResolver(CreateApplicationSchema),
@@ -81,10 +57,15 @@ export const ApplicationList = ({ applications, cohortAddress, builders, adminAd
         githubUsername: "",
       });
 
-      setIsSuccess(false);
+      if (onApplicationSuccess) {
+        onApplicationSuccess();
+      }
+
       router.refresh();
+      setIsSuccess(false);
     }
-  }, [isSuccess, form, router]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isSuccess, form]);
 
   useEffect(() => {
     const submitWithSignature = async () => {
@@ -104,6 +85,7 @@ export const ApplicationList = ({ applications, cohortAddress, builders, adminAd
           notification.error("Something went wrong");
           console.error("Error submitting application:", error);
         } finally {
+          router.refresh();
           setFormValues(null);
           setIsPending(false);
         }
@@ -126,83 +108,8 @@ export const ApplicationList = ({ applications, cohortAddress, builders, adminAd
     }
   };
 
-  const formatStatus = (status: string) => {
-    switch (status) {
-      case "PENDING":
-        return <span className="badge badge-warning">Pending</span>;
-      case "APPROVED":
-        return <span className="badge badge-success">Approved</span>;
-      case "REJECTED":
-        return <span className="badge badge-error">Rejected</span>;
-      default:
-        return <span className="badge">Unknown</span>;
-    }
-  };
-
-  const formatDate = (date: Date) => {
-    return new Date(date).toLocaleDateString();
-  };
-
-  if (!address) {
-    return <RainbowKitCustomConnectButton />;
-  }
-
   return (
-    <div className="w-full">
-      <h2 className="text-3xl mb-6 inline-block px-4 py-2 bg-primary text-secondary">Applications</h2>
-
-      {!address ? (
-        <div className="mb-4">Connect your wallet to view and submit applications</div>
-      ) : isBuilderorAdmin() ? (
-        <div className="mb-4">
-          <p className="text-sm">You are already member of this cohort. You cannot apply.</p>
-        </div>
-      ) : (
-        <>
-          {canApply() && (
-            <div className="mb-6">
-              <label
-                htmlFor="add-application-modal"
-                className="btn rounded-md btn-primary btn-sm font-normal space-x-2 normal-case"
-              >
-                Apply for Cohort
-                <Plus className="h-4 w-4" />
-              </label>
-            </div>
-          )}
-
-          {userApplications.length === 0 ? (
-            <p>You haven&apos;t submitted any applications yet.</p>
-          ) : (
-            <div className="space-y-4">
-              <h3 className="text-xl font-bold">Your Applications</h3>
-              {userApplications.map(application => (
-                <div key={application.id} className="border border-base-300 rounded-md p-4">
-                  <div className="flex justify-between items-start mb-2">
-                    <span className="text-sm ">{application.description}</span>
-                    {formatStatus(application.status)}
-                  </div>
-                  {application.githubUsername && (
-                    <p className="text-xs text-gray-400">
-                      GitHub:{" "}
-                      <a
-                        href={`https://github.com/${application.githubUsername}`}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="underline"
-                      >
-                        {application.githubUsername}
-                      </a>
-                    </p>
-                  )}
-                  <p className="text-xs text-gray-400 mt-2">Submitted on {formatDate(application.createdAt)}</p>
-                </div>
-              ))}
-            </div>
-          )}
-        </>
-      )}
-
+    <div>
       <input type="checkbox" id="add-application-modal" className="modal-toggle" />
       <label htmlFor="add-application-modal" className="modal cursor-pointer">
         <label className="modal-box relative bg-base-100 border border-primary ">
@@ -267,5 +174,3 @@ export const ApplicationList = ({ applications, cohortAddress, builders, adminAd
     </div>
   );
 };
-
-export default ApplicationList;
