+++ packages/nextjs/app/(routes)/cohort/[cohortAddress]/applications/page.tsx
@@ -1,4 +1,3 @@
-// app/cohort/[cohortAddress]/apply/page.tsx
 import React from "react";
 import Link from "next/link";
 import AdminApplicationList from "./_components/AdminApplicationList";
