+++ packages/nextjs/app/(routes)/cohort/[cohortAddress]/members/_components/EventsModal.tsx
@@ -46,7 +46,7 @@ export const EventsModal: React.FC<EventsModalProps> = ({
       <input type="checkbox" id="withdraw-events-modal" className="modal-toggle" checked={isOpen} readOnly />
       <label htmlFor="withdraw-events-modal" className="modal cursor-pointer " onClick={onClose}>
         <label
-          className="modal-box relative bg-base-100 max-w-4xl border border-primary bg-base-100"
+          className="modal-box relative bg-base-100 max-w-4xl border border-primary"
           onClick={e => e.stopPropagation()}
         >
           {/* dummy input to capture event onclick on modal box */}
