+++ packages/nextjs/app/(routes)/cohort/[cohortAddress]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 // app/cohort/[cohortAddress]/page.tsx
-import React, { useEffect, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { BuildersList } from "./_components/BuildersList";
@@ -83,21 +83,25 @@ const CohortPage = ({ params }: { params: { cohortAddress: string } }) => {
     router.push(`/cohort/${params.cohortAddress}/projects`);
   };
 
+  const fetchCohort = useCallback(async () => {
+    if (!params.cohortAddress) return;
+
+    try {
+      const response = await axios.get(`/api/cohort/${params.cohortAddress}`);
+      const cohort = response.data?.cohort;
+      setDbCohort(cohort);
+    } catch (error) {
+      console.error("Error fetching cohort from db:", error);
+    }
+  }, [params.cohortAddress]);
+
   useEffect(() => {
-    const fetchCohort = async () => {
-      if (!params.cohortAddress) return;
-
-      try {
-        const response = await axios.get(`/api/cohort/${params.cohortAddress}`);
-        const cohort = response.data?.cohort;
-        setDbCohort(cohort);
-      } catch (error) {
-        console.error("Error fetching cohort from db:", error);
-      }
-    };
-
     fetchCohort();
-  }, [params.cohortAddress, builderStreams]);
+  }, [fetchCohort, builderStreams]);
+
+  const handleApplicationSuccess = () => {
+    fetchCohort();
+  };
 
   return (
     <div className="max-w-4xl text-base-content">
@@ -144,6 +148,7 @@ const CohortPage = ({ params }: { params: { cohortAddress: string } }) => {
             dbBuilders={dbCohort?.Builder}
             dbAdminAddresses={dbCohort?.adminAddresses}
             applications={dbCohort?.Application}
+            onApplicationSuccess={handleApplicationSuccess}
           />
         </div>
       )}
